requirements.txt
# Core Data Processing
pandas==2.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.0

# Data Collection
yfinance==0.2.36
requests==2.31.0

# Visualization
streamlit==1.31.0
plotly==5.18.0

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
